Post-process generated site to use relative URIs
The goal of this commit is to make it possible to browse the generated
docusaurus site without a file server. Just open index.html and it
works. One benefit of this change is that it becomes possible to host
the generated site on Maven Central and browse old sites on static.javadoc.io.